Time: 0 ms (100.00%), Space: 6.6 MB (92.77%) - LeetHub

+++ 377-combination-sum-iv/377-combination-sum-iv.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {
-        vector<unsigned long> v(target+1, 0);
+        vector<unsigned int> v(target+1, 0);
         
         v[0] = 1;
-        for(unsigned long i=0; i < target+1; i++)
-            for(unsigned long x: nums)
+        for(unsigned int i=0; i < target+1; i++)
+            for(unsigned int x: nums)
                 if (x + i < target+1)
                     v[i+x] += v[i];
         return v[target];
